Refactor to be able to run it locally + added ruff for python formating
