+++ code-env/python/spec/requirements.dev.txt
@@ -1,7 +1,7 @@
 Flask>=0.9
-pandas>=0.23.0
+pandas==1.3.5
 numpy<1.24
 python-dotenv==0.19.0
 webaiku @ git+https://github.com/dataiku/solutions-contrib.git@main#egg=webaiku&subdirectory=bs-infra
-dataiku-api-client>=11.0.0
-https://design.solutions.dataiku-dss.io/public/packages/dataiku-internal-client.tar.gz
+dataiku-internal-client @ http://localhost:11200/public/packages/dataiku-internal-client.tar.gz
+ruff==0.4.5
