+++ python-lib/backend/fetch_api.py
@@ -1,27 +1,14 @@
-from flask import Blueprint, jsonify, request, send_file, current_app
-import pandas as pd
-
-is_local = False
-
-if not is_local:
-    from glm_handler.dku_model_trainer import DataikuMLTask
-    from glm_handler.dku_model_handler import ModelHandler
-    from glm_handler.dku_model_deployer import ModelDeployer
-    from glm_handler.glm_data_handler import GlmDataHandler
-    from glm_handler.dku_model_metrics import ModelMetricsCalculator
-    from backend.model_cache import setup_model_cache, update_model_cache
-    
-from backend.api_utils import format_models
-from backend.local_config import (dummy_models, dummy_variables, dummy_df_data,
-dummy_lift_data,dummy_get_updated_data, dummy_relativites, get_dummy_model_comparison_data, dummy_model_metrics)
-from backend.logging_settings import logger
 from io import BytesIO
 from time import time
-import traceback
+
 import dataiku
-
 import numpy as np
+import pandas as pd
+from flask import Blueprint, current_app, jsonify, request, send_file
+from glm_handler.dku_model_trainer import DataikuMLTask
 
+from backend.model_cache import update_model_cache
+from backend.utils.api_utils import format_models
 from backend.utils.dataiku_api import dataiku_api
 
 fetch_api = Blueprint("fetch_api", __name__, url_prefix="/api")
@@ -29,8 +16,6 @@ fetch_api = Blueprint("fetch_api", __name__, url_prefix="/api")
 
 @fetch_api.route("/models", methods=["GET"])
 def get_models():
-    if is_local:
-        return jsonify(dummy_models)
     if dataiku_api.global_dss_mltask is None:
         return jsonify({'error': 'ML task not initialized'}), 500
     try:
@@ -39,8 +24,6 @@ def get_models():
     except Exception as e:
         current_app.logger.exception("An error occurred while retrieving models")
         return jsonify({'error': str(e)}), 500
-    return jsonify(models)
-
 
 
 @fetch_api.route("/variables", methods=["POST"])
@@ -62,17 +45,10 @@ def get_variables():
         current_app.logger.error(f"An error occurred: {e}")
     
     return jsonify(variables)
-# local dev
-    # return jsonify(dummy_variables)
-
 
 
 @fetch_api.route("/data", methods=["POST"])
 def get_data():
-    if is_local:
-        import time
-        time.sleep(1)
-        return jsonify(dummy_df_data.to_dict('records'))
     try:
         current_app.logger.info("Received a new request for data prediction.")
         request_json = request.get_json()
@@ -92,8 +68,6 @@ def get_data():
 
 @fetch_api.route("/lift_data", methods=["POST"])
 def get_lift_data():
-    if is_local:
-        return jsonify(dummy_lift_data.to_dict('records'))
     current_app.logger.info("Received a new request for lift chart data.")
     request_json = request.get_json()
     full_model_id = request_json["id"]
@@ -102,20 +76,15 @@ def get_lift_data():
 
     current_app.logger.info(f"Model {full_model_id} is now the active version.")
     
-    print("-----> ", dataiku_api.model_cache)
     lift_chart = dataiku_api.model_cache[full_model_id].get('lift_chart_data')
     lift_chart.columns = ['Category', 'Value', 'observedAverage', 'fittedAverage']
     current_app.logger.info(f"Successfully generated predictions. Sample is {lift_chart.head()}")
     
     return jsonify(lift_chart.to_dict('records'))
-#     local dev
-    return jsonify(dummy_lift_data.to_dict('records'))
 
 
 @fetch_api.route("/update_bins", methods=["POST"])
 def get_updated_data(): 
-    if is_local:
-        return jsonify(dummy_get_updated_data.to_dict('records'))
     request_json = request.get_json()
 
     feature = request_json["feature"]
@@ -126,14 +95,10 @@ def get_updated_data():
     
     
     return jsonify(df.to_dict('records'))
-#     local dev
-    return jsonify(dummy_get_updated_data.to_dict('records'))
 
 
 @fetch_api.route("/relativities", methods=["POST"])
 def get_relativities():
-    if is_local:
-        return jsonify(dummy_relativites.to_dict('records'))
     request_json = request.get_json()
     full_model_id = request_json["id"]
     
@@ -142,29 +107,15 @@ def get_relativities():
     df.columns = ['variable', 'category', 'relativity']
     current_app.logger.info(f"relativites are {df.head()}")
     return jsonify(df.to_dict('records'))
-#     local dev
-    return jsonify(dummy_relativites.to_dict('records'))
+
 
 @fetch_api.route("/get_variable_level_stats", methods=["POST"])
 def get_variable_level_stats():
-    if is_local:
-        df = pd.DataFrame({'variable': ['VehBrand', 'VehBrand', 'VehBrand', 'VehPower', 'VehPower'], 
-                       'value': ['B1', 'B10', 'B12', 'Diesel', 'Regular'], 
-                       'coefficient': [0, 0.5, 0.32, 0, 0.0234],
-                       'standard_error': [0, 1.23, 1.74, 0, 0.9],
-                       'standard_error_pct': [0, 1.23, 1.74, 0, 0.9],
-                        'weight': [234, 87, 73, 122, 90], 
-                        'weight_pct': [60, 20, 20, 65, 35], 
-                        'relativity': [1, 1.23, 1.077, 1, 0.98]})
-        return jsonify(df.to_dict('records'))
-    print("variable level stats")
     request_json = request.get_json()
     full_model_id = request_json["id"]
-    
+
     current_app.logger.info(f"Model ID received: {full_model_id}")
 
-
-    print(dataiku_api.model_cache[full_model_id])
     df = dataiku_api.model_cache[full_model_id].get('variable_stats')
     df.columns = ['variable', 'value', 'relativity', 'coefficient', 'standard_error', 'standard_error_pct', 'weight', 'weight_pct']
     df.fillna(0, inplace=True)
@@ -172,16 +123,9 @@ def get_variable_level_stats():
     return jsonify(df.to_dict('records'))
 
 
-
-
 @fetch_api.route("/get_model_comparison_data", methods=["POST"])
 def get_model_comparison_data():
     start_time = time()
-    
-    if is_local:
-        df = get_dummy_model_comparison_data()
-        current_app.logger.info(f"Data fetched locally in {time() - start_time} seconds.")
-        return jsonify(df.to_dict('records'))
 
     current_app.logger.info("Received a new request for data prediction.")
     request_json = request.get_json()
@@ -208,21 +152,10 @@ def get_model_comparison_data():
     current_app.logger.info(f"Total execution time: {total_time} seconds.")
     return jsonify(merged_model_stats.to_dict('records'))
 
-   
-
-
 
 @fetch_api.route("/get_model_metrics", methods=["POST"])
 def get_model_metrics():
-    start_time = time()
-    
-    if is_local:
-        response_time = time() - start_time
-        print(f"Returned local dummy metrics in {response_time} seconds.")
-        return jsonify(dummy_model_metrics)
-    
     request_json = request.get_json()
-    print(request_json)
     
     model1, model2 = request_json["model1"], request_json["model2"]
     
@@ -246,54 +179,44 @@ def get_model_metrics():
     return jsonify(metrics)
 
 
-
-
 @fetch_api.route('/export_model', methods=['POST'])
 def export_model():
 
-    if is_local:
-        data = {'Name': ['John', 'Alice', 'Bob'], 'Age': [30, 25, 35]}
-        df = pd.DataFrame(data)
+    try:
+        request_json = request.get_json()
+        model = request_json.get("id")
+        if not model:
+            current_app.logger.error("error: Model ID not provided")
 
-        # Convert DataFrame to CSV format
-        csv_data = df.to_csv(index=False).encode('utf-8')
-    else:
-        try:
-            request_json = request.get_json()
-            print(request_json)
-            model = request_json.get("id")
-            if not model:
-                current_app.logger.error("error: Model ID not provided")
+        relativities_dict = dataiku_api.model_cache.get(model).get('relativities_dict')
+        if not relativities_dict:
+            current_app.logger.error("error: Model Cache not found for {model} cache only has {dataiku_api.model_cache.keys()}")
+        
+        current_app.logger.info(f"Relativities dict for model {model} is {relativities_dict}.")
+        
+        nb_col = (len(relativities_dict.keys()) - 1) * 3
+        variables = [col for col in relativities_dict.keys() if col != "base"]
+        variable_keys = {variable: list(relativities_dict[variable].keys()) for variable in variables}
+        max_len = max(len(variable_keys[variable]) for variable in variable_keys.keys())
 
-            relativities_dict = dataiku_api.model_cache.get(model).get('relativities_dict')
-            if not relativities_dict:
-                current_app.logger.error("error: Model Cache not found for {model} cache only has {dataiku_api.model_cache.keys()}")
-            
-            current_app.logger.info(f"Relativities dict for model {model} is {relativities_dict}.")
-            
-            nb_col = (len(relativities_dict.keys()) - 1) * 3
-            variables = [col for col in relativities_dict.keys() if col != "base"]
-            variable_keys = {variable: list(relativities_dict[variable].keys()) for variable in variables}
-            max_len = max(len(variable_keys[variable]) for variable in variable_keys.keys())
+        csv_output = ",,\n"
+        csv_output += "Base,,{}\n".format(relativities_dict['base']['base'])
+        csv_output += ",,\n" * 2
+        csv_output += ",,,".join(variables) + ",,\n" * 2
 
-            csv_output = ",,\n"
-            csv_output += "Base,,{}\n".format(relativities_dict['base']['base'])
-            csv_output += ",,\n" * 2
-            csv_output += ",,,".join(variables) + ",,\n" * 2
+        for i in range(max_len):
+            for variable in variables:
+                if i < len(variable_keys[variable]):
+                    value = variable_keys[variable][i]
+                    csv_output += "{},{},,".format(value, relativities_dict[variable][value])
+                else:
+                    csv_output += ",,,"
+            csv_output += "\n"
 
-            for i in range(max_len):
-                for variable in variables:
-                    if i < len(variable_keys[variable]):
-                        value = variable_keys[variable][i]
-                        csv_output += "{},{},,".format(value, relativities_dict[variable][value])
-                    else:
-                        csv_output += ",,,"
-                csv_output += "\n"
+        csv_data = csv_output.encode('utf-8')
 
-            csv_data = csv_output.encode('utf-8')
-
-        except KeyError as e:
-            current_app.logger.error(f"An error occurred: {str(e)}")
+    except KeyError as e:
+        current_app.logger.error(f"An error occurred: {str(e)}")
 
     csv_io = BytesIO(csv_data)
 
@@ -306,28 +229,16 @@ def export_model():
     )
 
 
-
 @fetch_api.route("/train_model", methods=["POST"])
 def train_model():
-    # Log the receipt of a new training request
-    if is_local:
-        import time
-        time.sleep(5)
-        return jsonify({'message': 'Model training initiated successfully.'}), 200
     
     request_json = request.get_json()
     current_app.logger.info(f"Received a model training request: {request_json}")
     
-    try: 
-        input_dataset = dataiku_api.webapp_config.get("training_dataset_string")
-        code_env_string = dataiku_api.webapp_config.get("code_env_string")
-        saved_model_id = dataiku_api.webapp_config.get("saved_model_id")
-        
-    except:
-        input_dataset = "claim_train"
-        code_env_string="py39_sol"
+    input_dataset = dataiku_api.webapp_config.get("training_dataset_string")
+    code_env_string = dataiku_api.webapp_config.get("code_env_string")
+    saved_model_id = dataiku_api.webapp_config.get("saved_model_id")
 
-        
     current_app.logger.info(f"Training Dataset name selected is: {input_dataset}") 
     
     distribution_function = request_json.get('model_parameters', {}).get('distribution_function')
@@ -369,25 +280,19 @@ def train_model():
         current_app.logger.info("Model training initiated successfully")
                 
         dataiku_api.model_cache = update_model_cache(dataiku_api.global_dss_mltask, dataiku_api.model_cache, dataiku_api.model_handler)
-            
-
         
         return jsonify({'message': 'Model training initiated successfully.'}), 200
     except Exception as e:
         current_app.logger.exception("An error occurred during model training")
         return jsonify({'error': str(e)}), 500
 
+
 @fetch_api.route("/get_dataset_columns", methods=["GET"])
 def get_dataset_columns():
     
     try:
-        # This try statement is just for local development remove the except 
-        # which explicitly assins the dataset name
-        try: 
-            dataset_name = dataiku_api.web_app_config.get("training_dataset_string")
-        except:
-            dataset_name = "claim_train"
-            
+        dataset_name = dataiku_api.web_app_config.get("training_dataset_string")
+        
         current_app.logger.info(f"Training Dataset name selected is: {dataset_name}")
 
         cols_dict = dataiku.Dataset(dataset_name).get_config().get('schema').get('columns')
