--- python-lib/backend/logging_settings.py
@@ -1,3 +1,0 @@
-import logging
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-logger = logging.getLogger(__name__)
