+++ python-lib/glm_handler/dku_relativities_handler.py
@@ -7,7 +7,6 @@ import numpy as np
 from dataiku import pandasutils as pdu
 from glm_handler.dku_utils import extract_active_fullModelId
 import logging
-from backend.logging_settings import logger
 
 class RelativitiesHandler:
     """
