+++ python-lib/glm_handler/glm_data_handler.py
@@ -7,8 +7,6 @@ import numpy as np
 from dataiku import pandasutils as pdu
 from glm_handler.dku_utils import extract_active_fullModelId
 
-from backend.logging_settings import logger
-
 
 class GlmDataHandler():
     def __init__(self):
